Keep track of scores and display winner of game
I wrote a function that will check for a winner of a round
that accepts the results of playround(). I also wrote a
function that will display the scores and a function that
will anounce a winner. I also added a function that will
check for game winner and call announceWinner() so that
the code will be clean. I then called each function from
inside the event listener, and deleted game() function.